more tests for the inverse renderer
This adds a test to the inverse renderer to reproduce the whole syntax
page. It's not a perfect test but should cover a lot of common cases.

The inverse renderer was adjusted to make the test pass. This might
actually have broken things because:

a) the test squashes whitespaces. but whitespaces are significant in
   a lot of DokuWiki's syntax
b) because some whitespace handling was changed it might not give
   desired results for the use in the table editor where line breaks
   need to be avoided