Fix linkwizard
This makes the link-wizard usable again.

It changes the handsontable.full.js, since this seems to be the only way
to get the full functionality of the link-wizard. (As suggested by
@lisps)
However the trade-of is, that we may have to redo this when we update
handsontable.
An alternative approach which does not change handsontable.full.js, but
aborts the cell-edit-session is implemented in
12251b3e6cb62bdb1b8f4ccb43169c6cfaf7eae7

WIKI-156, fixes #32