Use 0.34.0 of handsontable as basis for update
The .js file was created with hot-builder.

The following plugins were included/excluded:

  AutoColumnSize [included]
  AutoRowSize [included]
  Autofill [included]
  ColumnSorting [excluded]
  Comments [excluded]
  ContextMenu [included]
  CopyPaste [included]
  CustomBorders [excluded]
  DragToScroll [included]
  ManualColumnFreeze [included]
   └── Included dependency: ManualColumnMove
  ManualColumnMove [included]
  ManualColumnResize [included]
  ManualRowMove [included]
  ManualRowResize [included]
  MergeCells [included]
  MultipleSelectionHandles [included]
  ObserveChanges [excluded]
  PersistentState [excluded]
  Search [included]
  TouchScroll [excluded]
  UndoRedo [excluded]

We also only use the handsontable.css to excluded styles for the
unused pickaday datepicker. This should also prevent the problem from
issue #131.
The handsontable.css has also be created in the above mentioned process
and misses the styles for the excluded comment plugin